[BPK-4379] Migrate loading button examples
